requirements.txt
flask
flask-session
flask-sqlalchemy
redis
